+++ bocoel/models/adaptors/utils.py
@@ -1,6 +1,0 @@
-from collections.abc import Mapping, Sequence
-from typing import Any
-
-from bocoel.corpora import Storage
-
-collate = Storage.collate