+++ examples/ensemble.py
@@ -49,16 +49,12 @@ def main(
         "SetFit/sst2",
     ] = "SST2",
     ds_split: Literal["train", "validation", "test"] = "train",
-    idx: str = "idx",
-    sentence: str = "sentence",
-    label: str = "label",
-    sbert_model: str = "all-mpnet-base-v2",
     llm_model: str = "textattack/roberta-base-SST-2",
     batch_size: int = 16,
-    device: str = "cpu",
     sobol_steps: int = 20,
     index_threads: int = 8,
     optimizer_steps: int = 30,
+    device: str = "cpu",
     acqf: str = "ENTROPY",
     task: str = "EXPLORE",
     classification: Literal["logits", "classifier"] = "classifier",
@@ -66,43 +62,11 @@ def main(
 ) -> None:
     # The corpus part
     LOGGER.info("Loading datasets...", dataset=ds_path, split=ds_split)
-
     ds = datasets.load_dataset(ds_path)[ds_split]
     storage = DatasetsStorage(ds)
 
-    LOGGER.info(
-        "Creating embedder",
-        model=sbert_model,
-        device=device,
-    )
-    # embedder = SbertEmbedder(
-    #     model_name=sbert_model, device=device, batch_size=batch_size
-    # )
-
-    embedders = []
-
-    cuda_available = cuda.is_available()
-    device_count = cuda.device_count()
-    for i, model in enumerate(
-        [  # "textattack/bert-base-uncased-SST-2",
-            #   "textattack/roberta-base-SST-2",
-            #   "textattack/albert-base-v2-SST-2",
-            #   "textattack/xlnet-large-cased-SST-2",
-            #   "textattack/xlnet-base-cased-SST-2",
-            #   "textattack/facebook-bart-large-SST-2",
-            #   "textattack/distilbert-base-uncased-SST-2",
-            "textattack/distilbert-base-cased-SST-2"
-        ]
-    ):
-        # Auto cast devices
-        if cuda_available:
-            hf_device = f"cuda:{i%device_count}"
-        else:
-            hf_device = "cpu"
-        embedders.append(
-            HuggingfaceEmbedder(path=model, device=hf_device, batch_size=batch_size)
-        )
-    embedder = EnsembleEmbedder(embedders)
+    embedder = ensemble_embedder(batch_size=batch_size)
+    idx, sentence, label = idx_sentence_label(ds_path)
 
     LOGGER.info(
         "Creating corpus with storage and embedder",
@@ -110,11 +74,13 @@ def main(
         embedder=embedder,
         device=device,
     )
+
     corpus = ComposedCorpus.index_storage(
         storage=storage,
         embedder=embedder,
-        key=sentence,
+        keys=sentence.split(),
         index_backend=HnswlibIndex,
+        concat=" [SEP] ".join,
         distance=Distance.INNER_PRODUCT,
         threads=index_threads,
     )
@@ -147,9 +113,11 @@ def main(
     )
     adaptor: Adaptor
     if "setfit" in ds_path.lower():
-        adaptor = GlueAdaptor(num_texts=1 if ds_path == "SetFit/sst2" else 2)
+        adaptor = GlueAdaptor(idx, sentence, label)
     elif ds_path == "SST2":
-        adaptor = Sst2QuestionAnswer(idx=idx, sentence=sentence, label=label)
+        adaptor = Sst2QuestionAnswer(idx, sentence, label)
+    else:
+        raise ValueError(f"Unknown dataset {ds_path}")
 
     # ------------------------
     # The optimizer part.
@@ -209,6 +177,49 @@ def main(
 
     # Performs aggregation here.
     print("average:", np.average(scores))
+
+
+def idx_sentence_label(ds_path: str) -> tuple[str, str, str]:
+    if "setfit" in ds_path.lower():
+        idx = "idx"
+        sentence = "text" if ds_path == "SetFit/sst2" else "text1 text2"
+        label = "label"
+    elif ds_path == "SST2":
+        idx = "idx"
+        sentence = "sentence"
+        label = "label"
+    else:
+        raise ValueError(f"Unknown dataset {ds_path}")
+    return idx, sentence, label
+
+
+def ensemble_embedder(batch_size: int):
+    LOGGER.info("Creating embedder")
+    embedders = []
+
+    cuda_available = cuda.is_available()
+    device_count = cuda.device_count()
+    for i, model in enumerate(
+        [
+            # "textattack/bert-base-uncased-SST-2",
+            #   "textattack/roberta-base-SST-2",
+            #   "textattack/albert-base-v2-SST-2",
+            #   "textattack/xlnet-large-cased-SST-2",
+            #   "textattack/xlnet-base-cased-SST-2",
+            #   "textattack/facebook-bart-large-SST-2",
+            #   "textattack/distilbert-base-uncased-SST-2",
+            "textattack/distilbert-base-cased-SST-2"
+        ]
+    ):
+        # Auto cast devices
+        if cuda_available:
+            hf_device = f"cuda:{i%device_count}"
+        else:
+            hf_device = "cpu"
+        embedders.append(
+            HuggingfaceEmbedder(path=model, device=hf_device, batch_size=batch_size)
+        )
+    return EnsembleEmbedder(embedders)
 
 
 if __name__ == "__main__":